reworked long tests to run in sage
