+++ shell.nix
@@ -16,7 +16,7 @@ let
   };
 
   my-python = pkgs.python3;
-  python-with-my-packages = my-python.withPackages (p: with p; [
+  mypython = my-python.withPackages (p: with p; [
 	prettytable
 	scipy
 	sphinx
@@ -34,10 +34,17 @@ stdenv.mkDerivation {
   src = ./.;
 
   buildInputs = [ 
-  	python-with-my-packages
+    mypython
 	mysage
 	ripgrep
     nodePackages.pyright
 	tree
   ];
+
+  shellHook = ''
+    export PIP_PREFIX=$(pwd)/_build/pip_packages
+    export PYTHONPATH="$PIP_PREFIX/${mypython.sitePackages}:$PYTHONPATH"
+    export PATH="$PIP_PREFIX/bin:$PATH"
+    # unset SOURCE_DATE_EPOCH
+  '';
 }
